All library methods added to single file
