+++ cc_library/requirements.txt
@@ -1,3 +1,3 @@
 flake8
 black
-paho.mqtt.client
+paho.mqtt
