+++ cc_library/src/device/app.py
@@ -1,8 +1,26 @@
 from datetime import datetime
 import json
 
-from src.device.base_handler import BaseHandler
-from src.device.connection import Connection
+import paho.mqtt.client as mqtt
+
+
+def on_connect(self, userdata, flags, rc):
+    if rc == 0:
+        self.connected_flag = True  # set flag
+        print("connected OK")
+    else:
+        print("Bad connection Returned code=", rc)
+        self.bad_connection_flag = True
+
+
+def on_disconnect(client, userdata, rc):
+    print("disconnecting reason  " + str(rc))
+    client.connected_flag = False
+    client.disconnect_flag = True
+
+
+def on_log(client, userdata, level, buf):
+    print("log: ", buf)
 
 
 class App:
@@ -10,28 +28,24 @@ class App:
     Class App sets up the connection and the right handler
     """
 
-    connection_class = Connection
-    fallback_handler_class = BaseHandler
-
-    connection = None
-
     def __init__(self, config, device):
-        # self.handler_dict = handler_dict or handler_registry.copy()
         self.device = device
         self.config = json.load(config)
         self.name = self.config.get("id")
-
-        self.connection = self.connection_class(config=self.config, app=self)
-        # self.verbosity = int(self.config.get('verbosity', 1))
-        # self.api_url = config.get('api_url')
-        self.handler = BaseHandler
+        self.info = self.config.get("info")
+        self.host = self.config.get("host")
+        self.client = mqtt.Client(self.name)
+        self.client.on_message = self.on_message
+        self.client.on_log = on_log
+        self.client.on_connect = on_connect
+        self.client.on_disconnect = on_disconnect
 
     def start(self):
-        self.connection.connect()
+        self.connect()
 
     def subscribe_topic(self, topic):
         print("subscribed to topic", topic)
-        self.connection.client.subscribe(topic=topic)
+        self.client.subscribe(topic=topic)
 
     def send_status_message(self, msg):
         jsonmsg = {
@@ -43,4 +57,43 @@ class App:
             }
         }
         print(jsonmsg)
-        self.connection.client.publish("status", str(jsonmsg))
+        self.client.publish("status", str(jsonmsg))
+
+    def on_message(self, client, userdata, message):
+        print("message received ", str(message.payload.decode("utf-8")))
+        print("message topic=", message.topic)
+        self.handle(self, message)
+
+    def connect(self):
+        while True:
+            try:
+                self.client.connect(self.host, 1883, keepalive=60)
+                print("we zijn connected")
+                msg_dict = {
+                    "messageConnectionConfirmation": {
+                        "device_id": self.name,
+                        "time_sent":
+                            datetime.now().strftime("%Y-%m-%dT%H:%M:%S.%f%z"),
+                        "type": "connection",
+                        "message": {"connection": True},
+                    }
+                }
+                msg = json.dumps(msg_dict)
+                self.client.publish("connection", msg)
+                self.subscribe_topic("test")
+                self.client.loop_forever()
+                break
+            except (ConnectionError):
+                print("alles is kapot")
+
+    def handle(self, message):
+        message = message.payload.decode("utf-8")
+        message = json.loads(message)
+        message_type = message.get("messageInstructionTest").get("type")
+        if message_type == "instruction":
+            self.device.incoming_instruction(
+                message.get("messageInstructionTest").get("contents")
+            )
+        elif message_type == "status":
+            status = self.device.incoming_status()
+            self.send_status_message(status)
